fix: transform cast Int32Valer value to int64

+++ valer/valer.go
@@ -63,7 +63,7 @@ func (w Int32ValerWrapper) Value() (v driver.Value, err error) {
 		return
 	}
 	if p := w.Valer.Val(); p != nil {
-		v = *p
+		v = int64(*p)
 	}
 	return
 }
